requirements.txt
torchmetrics==1.4.2
timm==1.0.9
pytorch-lightning==2.2.0
pyyaml>=5.4.1
numpy>=1.21.0
onnx==1.16.2
onnxsim==0.4.36
tensorboard>=2.5.0
matplotlib==3.7.5
pycocotools==2.0.7
